requirements.txt
numpy
ase
mace-torch
rdkit
scikit-learn
umap-learn
tqdm